All dbalign features returns ESS alignments (-align option)
v0.8!!! -- almost there, in this stage, the program may be relased.
        -- The main missing feature is the selection of similarity
           matrix and the possible adition of the program to create
           the EC matrix

All the functions that align a bunch of ESSs contains the option
oscore (default True). If False, this option includes the aligned
ESSs in the outfile. Thus, the cmd -align option is conected with
the oscore option in the functions.

Beware, this parallel implementation mantain all the alignments in RAM,
so the computer may die if the database is large. This may be fixed
creating a sequential version of some functions that stores the
alignments on the run.